added the functionality to update and delete an order only if the customer owns

+++ app/Http/Controllers/Api/OrderController.php
@@ -4,7 +4,6 @@ namespace App\Http\Controllers\Api;
 
 use App\Http\Controllers\Controller;
 use App\Http\Requests\OrderRequest;
-use App\Models\Order;
 use App\Services\OrderService;
 use Illuminate\Http\Request;
 
@@ -15,14 +14,22 @@ class OrderController extends Controller
         $this->orderService = $orderService;
     }
 
-    public function index()
+    public function index(Request $request)
     {
-        $orders = $this->orderService->getOrders();
+        $request->validate([
+            'customer' => 'required|exists:customers,id',
+        ]);
+
+        $orders = $this->orderService->getOrders()->paginate();
         return response()->json($orders);
     }
 
-    public function show($order_id)
+    public function show(Request $request, $order_id)
     {
+        $request->validate([
+            'customer' => 'required|exists:customers,id',
+        ]);
+
         $order = $this->orderService->show($order_id);
         return response()->json($order);
     }
@@ -37,14 +44,31 @@ class OrderController extends Controller
 
     public function update(OrderRequest $request, $order_id)
     {
+        $request->validated();
         $data = $request->except('_token', '_method');
-        $this->orderService->update($data, $order_id);
-        return response()->json(['message' => 'Order updated successfully'], 200);
+        $orderExists = $this->orderService->getOrder($order_id) ?? false;
+        $owner = $orderExists->customer_id ?? false;
+
+        if ($orderExists && $owner == $request->customer) {
+            $this->orderService->update($data, $order_id);
+            return response()->json(['message' => 'Order updated successfully'], 200);
+        }
+        return response()->json(['message' => 'it was not possible to update the order'], 404);
     }
 
-    public function delete($order_id)
+    public function delete(Request $request, $order_id)
     {
-        $this->orderService->delete($order_id);
-        return response()->json(['message' => 'Order deleted successfully'], 200);
+        $request->validate([
+            'customer' => 'required|exists:customers,id',
+        ]);
+
+        $orderExists = $this->orderService->getOrder($order_id) ?? false;
+        $owner = $orderExists->customer_id ?? false;
+
+        if ($orderExists && $owner == $request->customer) {
+            $this->orderService->delete($order_id);
+            return response()->json(['message' => 'Order deleted successfully'], 200);
+        }
+        return response()->json(['message' => 'You are not owner of this order'], 403);
     }
 }
